Switch define_rust_fn to proc macro
This uses the newly stable proc macro support in Rust 1.30 to
reimplement `define_rust_fn` as a proc attribute macro. There's a few
benefits of this:

1. As this is now an attribute on normal Rust `fn`. This means that the
   full `fn` syntax is supported, `rustfmt` understands it, etc.

2. The unidiomatic `-> ()` is no longer required for unit functions.

3. We no longer need a weird descriptor value. We now use the
   function's name as the descriptor name.

4. Visibility like `pub` etc. works as expected instead of implicitly
   making everything `pub`

This will also be a good basis for defining native struct and function
types as well.

The downside of this is proc macro support is still very raw. We need to
lean on `syn` and `quote` heavily which are huge dependencies.