Basic support for converting thunks to callbacks
This can convert a boxed fun thunk to a callback via a dynamically
created adapter. This is the minimum to get callbacks working in the
REPL via `map`.

This causes LLVM crashes when lazily compiling functions that reference
constants across modules. It's not clear what's going on or if it's
directly related to the callback code.