Fix module ID generation, use eager compile
For some insane reason the LLVM Module IDs are in the same namespace as
JIT symbols on Linux. This means when we were building a module called
`sym?` with a function called `sym?` we would silently fail to look up
the function's address as the module was shadowing it.

Fix by using more verbose and unique module names. This allows us to
switch to eager compilation which seems to be less buggy in LLVM 7.
Specifically, this should work around two bugs:

- Using lazy compilation that refers to a global symbol in another
  module crashes. This prevented us from passing a callback that
  returned a constant box to `map`.

- Lazy compilation with an LLVM assert build crashes immediately on
  macOS. I would manually change to eager compilation on macOS to work
  around this when testing against assert builds.