Make non-entry JIT built funs private
These are accessory functions used by the entry point we're building.
Making them private means they can be discarded or optimised more
freely. This hacks around the fact we're building way too many of these
on each entry point.

Doing this now will also minimise the changes if we move to lazily
generating these functions inside code gen. At that point it will need
to set the linkage itself and won't know if it's a JIT or program built
fun.