Fix for referencing Rust functions from JIT
Every Rust function needs to be registered with the JIT so we can
resolve its entry point. Previously this was only done if we were JITing
a thunk to the Rust function itself. However, this would break if we
were JITing a function that indirectly referenced a Rust fun, e.g.:

`(fn () length)`

This could be made to work by "priming" the JIT by building a thunk to
`length` earlier.

Fix by moving the register to right beside where we reference the Rust
fun's entry point. This is a bit overzealous as it will also run in
non-JIT contexts. However, this should be harmless beyond a small
performance overhead.

This also cleans up some old misleading names and comments.