Promote BuiltReg to a MIR-level concept
Previously `BuiltReg` was a local hack in `build_reg`. It was used to
track the constness of values for the purposes of opportunistically
building constant lists instead of allocating them.

The responsibility of converting values to registers has been spread
across most of the MIR now that we support converting functions to
registers. This requires calling back in to `eval_hir` and its helpers
to build functions dynamically. As a result we were losing the constness
of regs when they left `build_reg`. For example, any list containing a
function wasn't considered const even if the function didn't capture any
values. This leads to unnecessary GC allocations at runtime.

This moves `BuiltReg` to `builder.rs` and converts the MIR to use that
as the primary way of tracking registers. This means the constness of
registers should be preserved while evaluating MIR. The built ops still
use raw `RegId` as the constness isn't useful to `codegen`.