Fix type tag check
We were using the `int_compare` path if any of the arguments could be an
Int. This is incorrectly for e.g. `(= Any Int)` and would cause an
invalid cast from `Any` to `Int`. This would usually evaluate to false
because random garbage would be read off the non-Int box but it's
clearly incorrect.