Fix spurious "not a fun" error with wrong ret type
When a function application has a known result type we feed that in to
the function expression itself via reverse type propagation. For
example, if `(foo)` must be an `Int` we want `foo` to be `(... -> Int)`.

When this fails we enter some heuristic code to produce a meaningful
type error for the user. The existing code was checking if the error is
due to a purity mismatch. If it's not, it assumes the expression is not
a function. This is clearly wrong in the case of having a return type
mismatch.

This changes the code to only declare that something is not a function
if it doesn't satisfy the topmost function type.