Fix building calls to panics
When we're building MIR (as opposed to evaluating it) any panic will
bubble all the way up and abort the compile. This causes even boxing a
function with a panic (conditional or not) to abort the compile.

Instead we should leave the panicing call in the built code. This
ensures the panic will only happen if it's reachable at runtime.