Pass the type of all expressions from inference
Previously we were passing the type from HIR inference to MIR in only
two places: conditionals and function application. This was because the
MIR needed type information in those cases for efficient code
generation. However, for things like type predicates that can be applied
to any expression we'd want the type information for all expression
kinds.

This moves `hir::Expr` in to `hir::ExprKind` and puts a span and result
type on `hir::Expr`. For lowered HIR the result type is `()` so it
doesn't need to be supplied.