Allow wildcard return types
Currently functions can either have no purity or return type, e.g.:
```
(fn () true)
```

Or both a purity and return type:
```
(fn () -> Bool)
```

This is  restrictive as sometimes we want to explicitly specify the
purity but want to infer the return type.

Add a specific hack to `lower_fun` that allows us to use `_` in place of
the return type when specifying the purity. This already had a
representation in HIR so it's isolated to lowering.