Track type information for variables in regs
We were previously deriving the possible types of a reg by its ABI type.
This is not as useful as I hoped; for example, every value loaded from a
rest list has the `Any` type which means don't have any useful type
information.

Instead of deriving the possible types of a reg as needed, explicitly
track in the reg value. This allows us to insert type information in two
places:

- `reg_to_value` already has the Arret type which we can use to limit
  the possible type tags as we build the `Value::Reg`

- Our HIR now tracks the types of all our nodes. We can use a new
  `value_with_arret_ty` function to annotate values as they pass through
  HIR nodes.

The result of this is equality and type predicates can now produce
optimised code in many more situations due to more precise type
information.