Add ArretFunId, Arret fun deduplication
Unlike Rust funs we don't have a stable identifier for Arret functions.
We need this for two reasons:

1. De-duplicating built Arret funs. Currently we build a new version
   every time we need a thunk or callback

2. Recursion detection. Right now we always recurse in to Arret funs but
   this prevents us from building functions with unbound recursion.

For a recursion detection heuristic a combination of span/source name
might be sufficient. However, for de-duplication we cannot have false
positives or else we could merge two unrelated functions.

This introduces a global ArretFunId and uses it for de-duplicating
functions. Recursion support is still unimplemented.