Initial support for out-of-line fun application
We currently always apply a fun inline which means unbounded function
applications cause a stack overflow at compile time. This adds initial
support for deciding if we should call a function out-of-line.

This currently just causes us to overflow the stack in a different
location when we continuously attempt to build the function itself. This
is because we don't realise we're already building the function and keep
trying to regenerate it.