Switch to two-phase private fun lowering
Currently when building a private fun (Rust, Arret or thunk->callback)
we build the fun and then allocate it an ID as we insert it to the
private fun Vec.

This falls down when dealing with recursive Arret funs. When
`id_for_arret_fun` goes to build the recursive call it can't "see" that
the function is already being built. This causes it to repeatedly build
the function over and over in an attempt to get its ID. Eventually the
compiler's stack is exhausted and it crashes.

Fix by using a counting ID type instead of an indexing ID type. This
lets us:

1. Allocate a new private fun ID

2. Insert it in to `arret_funs` for any recursive calls to see

3. Build the private fun

4. Insert it in to `private_funs`

This requires switching `private_funs` to be a `HashMap` which is less
efficient at runtime but requires little code churn.

This *still* doesn't allow us to compile recursive funs but now we crash
inside codegen.