Move planning allocs to the analysis phase
This allows us to determine if a private fun takes a task before we
actually build it. This allows us to fully determine its LLVM type with
data just from the analysis pass.

This is critical for recursive funs as we need to know the fun's type to
be able to call it. Eventually we can use this information to declare
all of the LLVM functions up front and then define them later.