Don't make task optional for private funs
We currently spend a lot of complexity checking if private funs require
a `task` parameter. This is due to `task` being a descendant of
Llambda's `world` parameter.

In Llambda functions not taking `world` could not allocate memory or
throw exceptions. This made not taking `world` a powerful optimisation
hint that was worth calculating. Arret is different in two important
ways:

1. We don't support arbitrary exceptions. The closest we come is
   `panic`ing the entire task and reporting a string. This does not
   require full unwinding support; the task is simply garbage collected
   entirely.

2. We don't support GC during task execution. This means we don't need
   to GC root temporaries when calling functions, `task` or not.

This means Arret does not care if a function takes `task` or not; it
behaves identically except for passing an additional parameter.

Additionally, all "public" functions (callbacks and thunks) need to take
a `task` for GC reasons anyway. This means we could only elide the
`task` for private functions. Since these have private linkage LLVM is
fully capable of detecting and removing the unused parameter.

One optimisation this prevents is not creating a `task` when a program
doesn't allocate memory or do I/O. While this is a cool optimisation
it's not worth the complexity of `task` tracking.