Always inline when evaluating a function body
We "cheat" and use our fun application logic to build a fun body.
However, this can easily lead to use generating a useless fun that just
tries to do an out-of-line call to itself.

Now that we have a concept of forcing inline application use that to
build fun bodies. This allows basic recursion to work.