Fix cond phi from wrong predecessor
When building a cond we were phi'ing from the first block of the branch.
This doesn't work if the branch has multiple blocks. Instead, use the
final block.

This also adds a compile test for recursion that requires a lot of the
changes over the past week to work correctly.