Support for Rust fun polymorphic purity
Currently all Rust funs with polymorphic purity are treated as impure.
While this is correct it's overly conservative and prevents certain
optimisations. For example, we can't evaluate `map` or `filter` at
compile time even if the lambda we pass is pure.

This is additionally tricky due to an LLVM restriction: the
`speculatable` attribute can only be applied for functions, not call
sites. This means we have to separately track the "upper bound" for a
functions purity in addition to the purity at every call site. Otherwise
we may falsely declare e.g. `map` as speculatable even though only some
of its calls are.