Add `eval` mode to our driver
This can be used to evaluate a program once, i.e. as a script. This is
essentially a mash up of the `compile` and `repl` code. It accepts
programs identical to `compile` but uses the REPL's evaluation engine.

At the moment this is intended to help with testing the compiler itself
as it allows running one-off test programs or individual `run-pass`
tests more quickly.