Split OpsABI in to PolymorphABI
`OpsABI` is intended for consumption by codegen and as such doesn't know
about closures, rest args, etc. These higher level concepts exist in the
MIR but aren't observable in the generated code.

However, the MIR was attempting to use `OpsABI` to track this higher
level ABI as well. To do so it tracked two additional pieces of
information:

- If the calling convention required passing a closure even if one does
  not exist. This is true for all callbacks and thunks at the moment.
  This was tracked in the obtusely named `OpsABI::external_call_conv`
  even though it was useless for codegen.

- If the final param was a rest argument. This was tracked by a
  `has_rest` passed wherever `wanted_abi` was passed.

This instead creates a `PolymorphABI` type which wraps the `OpsABI`
along with `has_closure` and `has_rest`.