Fix associated spans in RFI errors
RFI needs to load Arret types from the RFI libraries' metadata. As a
performance optimisation we avoid loading each RFI type in to our source
loader. Instead, we only load the type's string in to the source loader
when an error is actually encountered to allow the error to be
reported.

Currently we add the type string in to the source loader and then adjust
the span of the reported error. This works for trivial errors but
doesn't handle any associated span references, e.g. the start of the
list when a list is left open.

We can fix this by loading the type string with the offset of the
source file. Then, if we need to report an error we only need to load
the string as the span will already have the correct offset.