Don't consider funs capturing non-const vars const
Right now `value::to_const` will always return a value for Arret funs.
This is incorrect if the Arret fun captures non-const values; it will
always pass an empty closure which can result in undefined behaviour.

This checks if the Arret fun captures non-const values and refuses to
build a const value for it if it does.

This also adds a new `heap-alloc-count` helper for checking the number
of heap allocations a callback function makes. This can be used to
assert that we generate code that does not use the heap.