Use Err to handle divergence in eval_hir
Previously a divergent value was represented by a special `Value`
instance. Whenever this was encountered while evaling an expr evaluation
had to immediately stop and the divergent value passed back to the
caller. If this didn't occur then we were likely to build invalid ops
due to exiting in the middle of a block.

This was being fixed in a piecemeal fashion but it was easy to get
wrong. One observation is that the handling of divergent values is
almost identical to error values. This allows us to get divergence
handling "for free" without explicitly handling the case.

This adds a new `Error::Diverged` and converts a couple of places from
`Value` to `Result` where an error previously couldn't be handled.