Basic support for floating point syntax
This is implemented using a simple state machine that tracks which
component of the number we're in as we consume the input. The final
state is used to determine if we should parse as an `Int` or a `Float`.

This also allows `+` in front of integers for better EDN compatibility.
It comes for free when implementing `-`.