Make "keywords" self evaluating
In Clojure and Common Lisp keywords are similar to symbols but with a
`:` prefix. There are a number of differences between symbols and
keywords in Clojure but the most visible one is that keywords are
self-evaluating.

This makes symbols starting with `:` self evaluating. Unlike Clojure
these are still symbols and pass `sym?`; self-evaluation is the only
difference.