Speed up singleton equality
Our singleton boxes (true, false and ()) are equal iff their pointers
are equal. We can use this to speed up equality if either side is a
singleton type.

This is probably still no optimal - we're probably boxing values in some
cases we don't need to. However, we previously fell back to runtime
comparison which would box anyway so this is a strict improvement.