Fix is_a for polymorphic purity
For funs with a polymorphic return value we would use the parent top fun
to instantiate the sub fun. However, we would never consider the parent
top fun purity. This is inconsistent and caused code like this to not
type check:

```
(def _ (filter zero? '(0 1 2 3)))
```

This is because we didn't think `filter`'s polymorphic purity satisfies
the pure purity the context required.