Limit the size of the line we parse in the REPL
This sets a maximum parsed line length of 512. After that cutoff we'll
no longer provide parser hints or error highlighting.

This requires parsing the whole line and we don't support incremental
reparsing. This means in the worst case of pasting a large line
character-by-character we'll behave O(n!) with with the size of the
pasted line. This seems like a reasonable cutoff where a human isn't
typing the input.