Add slim REPL Docker image
The primary purpose of the Docker image is to access the REPL before
committing to setting up LLVM/Rust locally. It's not very practical for
actual compilation as it can only produce binaries for amd64 Linux
distros that resemble Debian buster.

This reduces the image size from 1.83GB to 130MB. This is much more
practical to upload to a Docker registry like Docker Hub.