Treat keywords as separate types during lowering
Right now both symbols and keywords are converted to `NsDatum::Ident`
before lowering. The lowering code then has a couple of half-baked
exceptions to deal with keywords as self-evaluating. However, in
destructuring and the fun position of application they are still treated
as symbol. There are likely more missing cases.

This changes lowering to convert keywords to `NsDatum::Keyword`. This
means they are uniformly handled as a different type of data than
symbols.