Don't consider polymorphics vars as literal types
This was presumably a copy-and-paste from `has_subtypes` where `true` is
a safe default. Polymorphic vars cannot be literal because any vars
bounded by literals will be converted to the type itself.

This was causing type inference to think this was always true:
```
(defn values-equal? #{T} ([left T] [right T]) -> Bool
  (= left right))
```