Test standalone 32bit compile
We currently run the unit tests using a 32bit target but nothing in the
tests actually asserts that the generated code is 32bit. This adds an
additional test that builds a binary, ensures it's 32bit and asserts its
output. Most of the test code is shared with the driver integration
tests.