Fix calling Arret funs with closures
We weren't accounting for the closure when saving/loading register
values. This causes is to either panic or generate invalid code.

This doesn't include a test because it requires recursion to trigger at
the moment. We will shortly be adding support for cost-based inlining
which will cause this to become well tested with our existing tests.