Implement cost-based inlining
This extends our inlining logic to use two new criteria:

1. The approximate cost of performing a call versus inlining. This is
   calculated by attempting both options and measuring the cost of the
   ops they build.

2. The amount of knowledge lost by calling through a closure and arg
   regs. This is referred to as the inlining preference factor. This is
   multiplied against the call cost.

Due to the amount of partial evaluation we support this is very eager to
inline. Out-of-line calls should only be used in extreme cases.

The cost/factor values used here are extremely approximate and not based
on any measurement. They can be tuned at a later date once our partial
evaluation and codegen is more mature.