Inline recursive funs while they make progress
Currently we abort inlining at the first recursive call. This is to
prevent partial unrolling of loop-like functions where we unroll 15
iterations of the function and then build a call for the 16th iteration.

This works by extending our apply cookie to contain a hash of our arg
list value. This lets us bail out of inlining early if it seems like
we're not making forward progress. Otherwise, we let inlining proceed up
until the inline limit.