Fix rfi-derive warnings on Rust Beta
Our macros internally convert what look like normal function names in to
a descriptor constant. This is an implementation detail; we want the
RFI consumers to use the original lowercase function name when
referencing the descriptor.

This triggers a new warning on Beta that the descriptor uses a lowercase
name. Silence it using `#[allow]`.