Merge compile-fail and pass tests
The core of these tests are very similar. This lets us unify the code
paths.

The ulterior motivation for this is to reduce memory usage when cross
compiling on Raspberry Pi. Right now we concurrently link the compiler,
`compile-fail` and `pass` at the same time which can balloon our memory
usage.