Initial type model rewrite
We have two first-class type references at the moment:

- `Poly` is a polymorphic type that can be or contain a type variable
  pointing to a generic type

- `Mono` is a monomorphic type that cannot contain type variables except
  for ones introduced in nested polymorphic funs

These are two different enums that implement some of the same traits.
However, this has lead to lots of either:

1. Code duplication inside our core type code to handle the two enum
   cases

2. Code that only handles one type or the other (typically only `Poly`)
   to avoid option 1

Change this by switching to a unified `ty::Ref` enum that looks like
`Poly`'s old structure. `Poly` and `Mono` then are used as marker types
on the `Var` variant to indicate if the `Var` is possible. `Mono` is a
never type so `Var` can't be constructed while `Poly` is a unit type.

This allows unification of our code paths in to what looks a lot like
the previous `Poly` case. When the code is specialised for `Mono` the
`Var` code paths should disappear because they're unreachable.

This commit is a nearly mechanical conversion to the new system. There
should be lots of further opportunity to simplify code such as `subst`
and make more code generic across `Ref` instead of hardcoding `Poly`.
These cleanups will follow.