Always inline self-executing funs
This spies on our strong reference count to determine if there are other
uses of a function expression. This has lots of false positives but
requires very little code and no new data structures.

Presumably once we can detect unused variables we'll have a proper
analysis pass we can hook in to.