Introduce RcId<T>
Currently we have global ID types for `tvars` and `pvars` that need to
be passed throughout the compiler whenever an operation may need to be
performed on a polymorphic type. In particular, nearly all of the `ty`
crate needs `tvars` to figure out polymorphic bounds.

This has two related downsides:

- Everyone that descends through HIR needs to accumulate a set of
  `tvars` as it walks the expression tree. In the case of `infer`
  there's an entire `FunCtx` struct for that purpose. This is both easy
  to get wrong (resulting in crashes) and expensive (due to hash
  operations and source name clones).

- Code that only incidentally deals with types such as `eval_hir` has a
  limited number of type operations it can perform due to not wanting
  the overhead of the previous point. Due to how our type system works
  this is required even for mono types.

This creates a new `RcId<T>` type, a reference-counted pointer that uses
pointer identity Traits such as `Hash`, `Eq`, `Ord` etc. are implemented
in terms of the value's memory location.  This means that the value
returned by `RcId::new()` is considered equal to itself and its clones
regardless of the value it points to.

This causes cloned the IDs to become more expensive but should overall
be much cheaper due to not having to maintain `tvars` and `pvars` maps.