Make ty_ref_is_a return a `bool`
The code in `is_a.rs` is some of the most complex and dense in the
compiler. One of the main reasons is it doesn't just query if a type
definitely satisfies another type; it also queries if any value of that
type may satisfy the other type. This is very complex to determine and
means we miss a lot of opportunities to bail out of the type test early
during runtime.

This "may" result is only treated separately from "no" in one case:
inferring type predicates. However, type predicates are a small subset
of all types and don't require the full type test logic.

This makes `ty_ref_is_a` return a `bool` and adds a new
`pred::TestTy::match_subject_ref` that supports a ternary result. This
results in some code and responsibility duplication but the result is
much easier to follow and should have better runtime performance.

At the same time this adds some `From` implementations for `ty::Map` and
`ty::List` that mirror the ones already on `ty::Fun` and `ty::TopFun`.