Partial Sym codegen support
We have two types of interned symbols:

1. Inlined symbols can be packed in to our 8 byte representation
2. Indexed symbols reference a long name in our `Interner`. In the
   future we'll also support global indexed symbols that are shared
   across a compiled program.

When we're running in a JIT we can't use global indexed symbols as we
don't want to maintain a single static global table across threads.
However, we also can't use the current `Interner`-indexed symbols as
generated code can't act as a GC root. As a consequence the next garbage
collection would invalidate all the constant indices in our generated
code.

Instead, we introduce a new `intern_static` which allows us to intern
symbols that are never garbage collected. There is some attempt at
optimisation but this leaks all symbols that were in the `Interner` at
the time.

In the non-JIT case we allow creating inline symbols but panic when
trying to create what should be a globally indexed one. This is because
we don't support generating or using the global index data structures
yet.