Support for global interned symbols
This creates a global interned names array when we generate programs.
This is then passed to `arret_runtime_launch_task` which then stores it
on the `Interner`.

At the moment this panics if we try to intern a symbol at runtime. We
build an array to map global indexes to names, but not a hash table to
map back. We panic to prevent ourselves from creating a local indexed
symbol that doesn't have the same `u64` value as the global one. This
breaks invariants we depend on for testing equality etc.