Add intrinsic (float) for ints
This also splits or intrinsics in to "eval" and "build" phases. The vast
majority of intrinsics can't do a better than evaluating to a static
constant value at compile time.

Instead of having every "eval" implementation duplicate the constant
case we can have a second phase of intrinsics that runs if we can't
const eval.