Basic math intrinsics
This generates MIR ops for `+`, `-`, `*` and `/` when all operands have
a definite type. The integer ops include "Checked" in their name but
this is entirely aspirational at the moment.

This also adds a `SizedListIterator` which provides more traditional
iterator interface around lists of a known size. This is needed for our
math intrinsics and presumably other intrinsics with variable arity in
the future.