Add (fold) and fix select the poly fun evidence
Our `(fold)` corresponds to the three parameter version of `(reduce)` in
Clojure. The implementation directly maps to a Rust `fold`.

Unfortunately this ran in to a bug with selection with a polymorphic fun
as evidence. We were naively using the evidence function's return type
(e.g. `N` for `+`) as the evidence for the outer fun. This meant the
evidence polymorphic type "leaked" to the selecting fun's selected type
before it was introduced. This causes an ICE.

We have two alternatives for dealing with polymorphic evidence funs:

1. Do a recursive selection where we pass the known types from the
   target fun in to the evidence fun and use that to calculate the
   return type. This is possible but complex.

2. Do nothing and depend on the fact the target fun is probably already
   polymorphic and expresses the type relationship we care about. This
   is the option implemented in this commit.