Speed up `into_poly_ty_args`
Right now we're cloning all of the IDs and variables and then adding any
bounds as needed. This is very functional but also does a lot of
allocations.

Instead, pre-reserve the needed size of `tvar_types` and `pvar_purities`
when initialising and then add any missing vars when `into_poly_ty_args`
is called. In the common case this will actually do nothing and just
return the var `HashMap`s directly.