Represent fixed lists with never rests
As an experiment I attempted to raise an error whenever our type
inference couldn't select on a type or purity var. This was to both make
our type system more strict and catch corner cases in our select code.
However, this ran in to a number of errors like the following:

```
error: cannot determine type of type variable `T` in this context
  --> ./tests/eval-pass/list.arret:17:19
   |
17 |   (assert-eq '(1) (cons 1 '()))
   |                   ^^^^^^^^^^^^
```

This was a previously known (and mostly forgotten) issue where an empty
list can't be used as type evidence against another list. In this case
the result type becomes `(Listof Int Any ...)` because it defaults to
`T` being `Any`.

It may be possible to hack this in to `select.rs` and `subst.rs` without
changing our list representation. However, this also makes `is_a`,
`unify` and `infer` simpler as they can use normal type system logic on
rest types without special casing if the rest is present or not.