Split select's poly evidence fun check
PR #51 fixed a crash when using polymorphic funs as evidence during type
selection. However, it was a bit too aggressive in that it stopped
collecting purity evidence as well. This was highlighted during an
experimental to require all purity and type vars to be selected during
type inference. This is a representative error:

```
error: cannot determine type of purity variable `->_` in this context
  --> ./tests/eval-pass/list.arret:43:16
   |
43 |   (assert-eq 7 (fold + 1 '(1 2 3)))
   |                ^^^^^^^^^^^^^^^^^^^
```

Because `+` is polymorphic on its return type we didn't use its purity
as evidence.

Fix by treating type and purity as independent dimensions when
determining which part of the fun type we can use as evidence.