Switch infer errors from strings to ty::Ref
Before our big `TVarId`/`PVarId` refactor it wasn't possible to render a
type as a string without knowing the full polymorphic var context on the
error. This forced us to render the types at the site of the inference
error and store it as a `Box<str>`. This led to a lot of gross error
presentation code being mixed with the type inference core.

This switches to just using `ty::Ref<ty::Poly>` which we can now render
without any external information. All of the error rendering now lives
inside `error.rs`