Support rest args in polymorph type selection
When building a polymorph of an Arret fun we try to "seed" the
polymorphic types with the ABI of the polymorph. This is to support more
precise monomorphisation of the Arret fun - for example, we can generate
better code if we know a value is an `Int` instead of just a `Num`.

The existing code didn't want to deal with rest args so it did some
gross zipping thing. Instead make a proper method on `PolymorphABI` and
use the full param list type. This allows us to collect type information
from our rest arg.

This depends on the rest arg always being in a `BoxedABIType::List`.
This currently holds true and we'll panic if that changes.