Significantly simplify macro system
The current macro system is a slightly simplified form of the one in
`etaoins/llambda`. By default symbols represent macro variables in the
pattern and template. However, if the symbols are designated as macro
literals they're instead matched during macro expansion. Bound literal
identifiers are matched based on their binding while unbound identifiers
are matched on their name.

This ultimately comes from R7RS, particularly this text:

```
An element in the input matches a literal identifier if and only if it
is an identifier and either both its occurrence in the macro expression
and its occurrence in the macro definition have the same lexical
binding, or the two identifiers are the same and both have no lexical
binding.
```

While this very literal implementation of R7RS worked as expected, it
had a number of disadvantages:

1. Simple bound values such as primitives and variables have trivial
   equality checks. However, as we add increasingly abstract and complex
   bound values this becomes harder to provide. For example, we attempt
   to match on bound types but `(U Int Float)` and `(U Float Int)` are
   two unequal types that are otherwise indistinguishable by the type
   system.

   We had to provide `Eq` and `Hash` implementations on types to support
   macros even though those operations had tenuous applicability in
   general. While it was a corner case to macro match on types the
   existence of the `Eq` and `Hash` implementations made it tempting to
   use them in more dangerous situations.

2. The macro code consistently shows up at the top of our profiling
   runs. This is due to how macro-heavy our language is (e.g. almost
   everything is inside a `defn` macro) and that macro expansion must
   unconditionally happen over all code to determine its structure. The
   `MacroVar` abstraction made it difficult to control the number of
   scope lookups and allocations we performed.

Based on a quick survey of the macros defined in Clojure this system can
be vastly simplified. Literal identifiers are rarely used in Clojure
macro patterns, instead depending on parameter ordering or the richer
data types available to Clojure. When identifiers are used they're nearly
always keywords (e.g. `cond`'s use of `:else`).

This allows us to leverage the distinction between symbols and macros in
Clojure. We can treat symbols always as variables while keywords can be
treated always at literals. This removes the concept of macro-designated
literals entirely.

As keywords cannot map to a bound value by definition this removes the
`MacroVar` abstraction entirely in favour of `ns::Ident`. This is both
less complex and should be significantly more performant.

This also remove the `Eq` and `Hash` impls for our type system value
object now they're unused.

While this sounds like a big conceptual shift it actually did not
affect any of our macros outside or our macro integration tests. That
is further evidence that our complex system of literals wasn't well used
and the keyword literals should hopefully meet our future requirements.

When running our integration tests inside Callgrind this resulted in
3.7% fewer memory allocations and 20% fewer instructions spent in
`scope.rs`